backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
pydantic-settings
thefuzz
python-Levenshtein
azure-storage-blob
azure-cosmos
